+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to the "aiXcoder" extension will be documented in this file.
 
 ## [0.1.3] - %DATE%
 ### Added
+- Added a configuration to set the prefered position of long results.
 
 ### Changed